Code clean up

+++ src/main/java/ru/itmo/lab2/Calculator.java
@@ -41,12 +41,12 @@ class Calculator {
         return number1 * number2;
     }
 
-    public static int divide(int number1, int number2) {
-        return number1 / number2;
+    public static double divide(int number1, int number2) {
+        return (double) number1 / number2;
     }
 
-    public static long divide(long number1, long number2) {
-        return number1 / number2;
+    public static double divide(long number1, long number2) {
+        return (double) number1 / number2;
     }
 
     public static double divide(double number1, double number2) {
